requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-core
openai
python-dotenv
pydantic
langchain-text-splitters
pydantic-settings 
pytest
httpx
pytest-cov 
pytest-asyncio